+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/params/params-object.ts
@@ -1,3 +1,4 @@
+
 import { zodToOpenAPI } from '../../../../../helper/zod-to-openapi.js'
 import type { Parameters } from '../../../../../openapi/index.js'
 import { getToSafeIdentifier } from '../../../../../utils/index.js'
